requirement.txt
pandas
numpy
scikit-learn
xgboost
joblib
requests
beautifulsoup4
feedparser
python-dotenv